+++ unfucker.user.js
@@ -49,14 +49,14 @@ const main = async function () {
     disableTumblrLive: { type: "checkbox", value: "checked" },
     disableTumblrDomains: { type: "checkbox", value: "checked" },
     revertActivityFeedRedesign: { type: "checkbox", value: "checked" },
-    revertMessagingRedesign: { type: "checkbox", value: "checked" },
     revertSearchbarRedesign: { type: "checkbox", value: "checked" },
     enableCustomTabs: { type: "checkbox", value: "" },
     enableReblogPolls: { type: "checkbox", value: "" },
     disableTagNag: { type: "checkbox", value: "checked" },
     reAddHomeNotifications: { type: "checkbox", value: "checked" },
     displayVoteCounts: { type: "checkbox", value: "" },
-    showNsfwPosts: { type: "checkbox", value: "" }
+    showNsfwPosts: { type: "checkbox", value: "" },
+    disableScrollingAvatars: { type: "checkbox", value: ""}
   };
   let pathname = location.pathname.split("/")[1];
   const $a = selector => document.querySelectorAll(selector);
@@ -232,7 +232,6 @@ const main = async function () {
     {"name": "liveSteamingWebPayments", "value": !configPreferences.disableTumblrLive.value},
     {"name": "domainsSettings", "value": !configPreferences.disableTumblrDomains.value},
     {"name": "activityRedesignM3", "value": !configPreferences.revertActivityFeedRedesign.value},
-    {"name": "messagingRedesign", "value": !configPreferences.revertMessagingRedesign.value},
     {"name": "improvedSearchTypeahead", "value": !configPreferences.revertSearchbarRedesign.value},
     {"name": "configurableTabbedDash", "value": configPreferences.enableCustomTabs.value?true:false},
     {"name": "allowAddingPollsToReblogs", "value": configPreferences.enableReblogPolls.value?true:false},
@@ -706,6 +705,10 @@ const main = async function () {
               $a(".pollDetailed").forEach(elem => elem.classList.remove("pollDetailed"));
             }
             break;
+          case "__disableScrollingAvatars":
+            if (value) document.getElementById("__as").innerText = `${keyToCss("stickyContainer")} > ${keyToCss("avatar")} { position: static !important; }`;
+            else document.getElementById("__as").innerText = "";
+            break;
         };
       };
       const rangeEvent = (id, value) => {
@@ -852,9 +855,15 @@ const main = async function () {
                 <label for="__showFollowingLabel">Toggle</label>
               </li>
               <li>
+<<<<<<< HEAD
                 <span>display exact vote counts on poll answers</span>
                 <input class="configInput" type="checkbox" id="__displayVoteCounts" name="displayVoteCounts" ${configPreferences.displayVoteCounts.value}>
                 <label for="__displayVoteCounts">Toggle</label>
+=======
+                <span>disable avatars scrolling with posts</span>
+                <input class="configInput" type="checkbox" id="__disableScrollingAvatars" name="disableScrollingAvatars" ${configPreferences.disableScrollingAvatars.value}>
+                <label for="__disableScrollingAvatars">Toggle</label>
+>>>>>>> 706b67f516782cdba9ab6643f078e1e0094deaee
               </li>
               <li>
                 <span>content positioning</span>
@@ -908,11 +917,6 @@ const main = async function () {
                 <span>revert activity feed redesign</span>
                 <input class="configInput" type="checkbox" id="__revertActivityFeedRedesign" name="revertActivityFeedRedesign" ${configPreferences.revertActivityFeedRedesign.value}>
                 <label for="__revertActivityFeedRedesign">Toggle</label>
-              </li>
-              <li>
-                <span>revert messaging redesign</span>
-                <input class="configInput" type="checkbox" id="__revertMessagingRedesign" name="revertMessagingRedesign" ${configPreferences.revertMessagingRedesign.value}>
-                <label for="__revertMessagingRedesign">Toggle</label>
               </li>
               <li>
                 <span>revert searchbar update</span>
@@ -1011,7 +1015,13 @@ const main = async function () {
             height: calc(450px * ${configPreferences.messagingScale.value});
           }
         `;
-
+        const avatarStyle = document.createElement("style");
+        avatarStyle.id = "__as";
+        document.head.append(avatarStyle);
+        if (configPreferences.disableScrollingAvatars.value) {
+          avatarStyle.innerText = `${keyToCss("stickyContainer")} > ${keyToCss("avatar")} { position: static !important; }`;
+        }
+        
         observer.observe(target, { childList: true, subtree: true });
       };
       const setupButtons = () => {