+++ changelog.md
@@ -1,9 +1,15 @@
 # Changelog
 
+<<<<<<< HEAD
 ## v5.3.0
 - Overhauled mutuation functions
 - Moved the "display exact vote counts on poll answers" feature to the "general configuration" category
 - Removed the "display full note count" feature since tumblr no longer shortens post note counts
+=======
+## v5.2.2
+- Added the "disable avatars scrolling with posts" feature
+- Removed the "revert messaging redesign feature" (no longer functioned properly)
+>>>>>>> 706b67f516782cdba9ab6643f078e1e0094deaee
 
 ## v5.2.1
 - Fixed a minor text colour discrepancy with some palettes